Pickup new buildpack version from go.mod (#33)
This will help ensure we are testing the same version of a buildpack
in the go mod and the builder. Moreover, since dependabot will send PRs
for go.mod updates, we can not automatically test the release buildpack
of the same version.

Adding github.com/cloudfoundry/build-system-cnb as a faux go dependency
so it will also be managed by dependabot.